Add files via upload
In this modern era of urbanizations, a road traffic accident is the ninth leading cause of death and the first leading cause of young people (15-29 years old) death. Every year 1.25 million people have died in the road accident and 50 million people suffered from road accidents. The accidents and the fatalities are varied from country to country. The lower and middle-income countries which own only 54% of world vehicles, 84% of total accidents and 90% of total fatalities have occurred in those countries. Bangladesh is one of the lower middle-income country. Every year around 4000 people died and around 9000 people are suffered by the road accident in this country.

Dhaka-Bagura is one of the main portions of highways of this country. Most of the accident happened in rainy season (July to September). 76.6% of accidents happened in the day times (4 am-5 pm). 23.4% accident occurs in the whole night. 55.10% of these accidents occur from 6 pm to 9 pm. 69% accident occurs in no-junction area and 66.97% of that are fatal accident. There are 52.54% pedestrian hit accidents followed by head on and rear end which are 18.89% and 11.06%. Buses are responsible for 39% of accidents, followed by trucks 25% and 2-3-wheeler and non-motorized vehicles 13%. The black spot which scatters in only 12 km of road section found by GIS analysis, has 40.26% of total collusion. Among them, 57.14% are pedestrian hit accident followed by 14.29% head-on collision. Buses and trucks both of these are liable to 72% of accidents in the black spot. 44% of these are bus followed by the truck which is 28%. The non-motorized and 2/3-wheeler were involved in mentionable 12% of total accident in this section. Therefore, it can be said that road accident has become one of the major problems in Bangladesh. The necessary steps should be taken to improve the road safety conditions of the country as soon as possible.